+++ src/Typin/Typin/Internal/Exceptions/CommandResolverExceptions.cs
@@ -3,7 +3,6 @@
     using System;
     using System.Collections.Generic;
     using Typin.Attributes;
-    using Typin.AutoCompletion;
     using Typin.Exceptions;
     using Typin.Internal.Extensions;
     using Typin.Schemas;
@@ -11,22 +10,11 @@
     /// <summary>
     /// Internal exceptions. Provide more diagnostic information here.
     /// </summary>
-    internal static class ResolversExceptions
+    internal static class CommandResolverExceptions
     {
-        public static TypinException InvalidSupportedModesInDirective(Type type)
-        {
-            var message = $@"
-Directive '{type.FullName}' contains an invalid mode in SupportedModes parameter.
-Either the type does not implement {nameof(ICliMode)} or CLI mode was not registered.
-
-If you're experiencing problems, please refer to the readme for a quickstart example.";
-
-            return new TypinException(message.Trim());
-        }
-
         public static TypinException InvalidSupportedModesInCommand(Type type)
         {
-            var message = $@"
+            string message = $@"
 Command '{type.FullName}' contains an invalid mode in SupportedModes parameter.
 Either the type does not implement {nameof(ICliMode)} or CLI mode was not registered.
 
@@ -35,17 +23,9 @@ If you're experiencing problems, please refer to the readme for a quickstart exa
             return new TypinException(message.Trim());
         }
 
-        public static TypinException DuplicatedShortcut(ShortcutDefinition definition)
-        {
-            var message = $@"
-Shortcut '{definition.Modifiers}+{definition.Key}' is already used by Typin.";
-
-            return new TypinException(message.Trim());
-        }
-
         public static TypinException InvalidCommandType(Type type)
         {
-            var message = $@"
+            string message = $@"
 Command '{type.FullName}' is not a valid command type.
 
 In order to be a valid command type, it must:
@@ -58,24 +38,9 @@ If you're experiencing problems, please refer to the readme for a quickstart exa
             return new TypinException(message.Trim());
         }
 
-        public static TypinException InvalidDirectiveType(Type type)
-        {
-            var message = $@"
-Directive '{type.FullName}' is not a valid directive type.
-
-In order to be a valid directive type, it must:
-- Not be an abstract class
-- Implement {typeof(IDirective).FullName}
-- Be annotated with {typeof(DirectiveAttribute).FullName}
-
-If you're experiencing problems, please refer to the readme for a quickstart example.";
-
-            return new TypinException(message.Trim());
-        }
-
         public static TypinException NoCommandsDefined()
         {
-            var message = $@"
+            string message = $@"
 There are no commands configured in the application.
 
 To fix this, ensure that at least one command is added through one of the methods on {nameof(CliApplicationBuilder)}.
@@ -86,7 +51,7 @@ If you're experiencing problems, please refer to the readme for a quickstart exa
 
         public static TypinException TooManyDefaultCommands()
         {
-            var message = $@"
+            string message = $@"
 Application configuration is invalid because there are too many default commands.
 
 There can only be one default command (i.e. command with no name) in an application.
@@ -97,7 +62,7 @@ Other commands must have unique non-empty names that identify them.";
 
         public static TypinException CommandsWithSameName(string name, IReadOnlyList<CommandSchema> invalidCommands)
         {
-            var message = $@"
+            string message = $@"
 Application configuration is invalid because there are {invalidCommands.Count} commands with the same name ('{name}'):
 {invalidCommands.JoinToString(Environment.NewLine)}
 
@@ -107,35 +72,9 @@ Names are not case-sensitive.";
             return new TypinException(message.Trim());
         }
 
-        public static TypinException DirectiveWithSameName(string name, IReadOnlyList<DirectiveSchema> invalidDirectives)
+        public static TypinException ParametersWithSameOrder(CommandSchema command, int order, IReadOnlyList<CommandParameterSchema> invalidParameters)
         {
-            var message = $@"
-Application configuration is invalid because there are {invalidDirectives.Count} directives with the same name ('[{name}]'):
-{invalidDirectives.JoinToString(Environment.NewLine)}
-
-Directives must have unique names.
-Names are not case-sensitive.";
-
-            return new TypinException(message.Trim());
-        }
-
-        public static TypinException DirectiveNameIsInvalid(string name, Type type)
-        {
-            var message = $@"
-Directive '{type.FullName}' is invalid because its name is empty or whitespace ('[{name}]').
-
-Directives must have unique, non-empty, and non-whitespace names.
-Names are not case-sensitive.";
-
-            return new TypinException(message.Trim());
-        }
-
-        public static TypinException ParametersWithSameOrder(
-            CommandSchema command,
-            int order,
-            IReadOnlyList<CommandParameterSchema> invalidParameters)
-        {
-            var message = $@"
+            string message = $@"
 Command '{command.Type.FullName}' is invalid because it contains {invalidParameters.Count} parameters with the same order ({order}):
 {invalidParameters.JoinToString(Environment.NewLine)}
 
@@ -144,12 +83,9 @@ Parameters must have unique order.";
             return new TypinException(message.Trim());
         }
 
-        public static TypinException ParametersWithSameName(
-            CommandSchema command,
-            string name,
-            IReadOnlyList<CommandParameterSchema> invalidParameters)
+        public static TypinException ParametersWithSameName(CommandSchema command, string name, IReadOnlyList<CommandParameterSchema> invalidParameters)
         {
-            var message = $@"
+            string message = $@"
 Command '{command.Type.FullName}' is invalid because it contains {invalidParameters.Count} parameters with the same name ('{name}'):
 {invalidParameters.JoinToString(Environment.NewLine)}
 
@@ -161,7 +97,7 @@ Names are not case-sensitive.";
 
         public static TypinException TooManyNonScalarParameters(CommandSchema command, IReadOnlyList<CommandParameterSchema> invalidParameters)
         {
-            var message = $@"
+            string message = $@"
 Command '{command.Type.FullName}' is invalid because it contains {invalidParameters.Count} non-scalar parameters:
 {invalidParameters.JoinToString(Environment.NewLine)}
 
@@ -175,7 +111,7 @@ If it's not feasible to fit into these constraints, consider using options inste
 
         public static TypinException NonLastNonScalarParameter(CommandSchema command, CommandParameterSchema invalidParameter)
         {
-            var message = $@"
+            string message = $@"
 Command '{command.Type.FullName}' is invalid because it contains a non-scalar parameter which is not the last in order:
 {invalidParameter}
 
@@ -187,43 +123,9 @@ If it's not feasible to fit into these constraints, consider using options inste
             return new TypinException(message.Trim());
         }
 
-        public static TypinException OptionsWithDigitStartingName(CommandSchema command, IReadOnlyList<CommandOptionSchema> invalidOptions)
-        {
-            var message = $@"
-Command '{command.Type.FullName}' is invalid because it contains one or more options with a name starting from digit or short names with a digit:
-{invalidOptions.JoinToString(Environment.NewLine)}
-
-Options must have a name starting from char other than digit, while short name must not be a digit.";
-
-            return new TypinException(message.Trim());
-        }
-
-        public static TypinException OptionsWithNoName(CommandSchema command, IReadOnlyList<CommandOptionSchema> invalidOptions)
-        {
-            var message = $@"
-Command '{command.Type.FullName}' is invalid because it contains one or more options without a name:
-{invalidOptions.JoinToString(Environment.NewLine)}
-
-Options must have either a name or a short name or both.";
-
-            return new TypinException(message.Trim());
-        }
-
-        public static TypinException OptionsWithInvalidLengthName(CommandSchema command, IReadOnlyList<CommandOptionSchema> invalidOptions)
-        {
-            var message = $@"
-Command '{command.Type.FullName}' is invalid because it contains one or more options whose names are too short:
-{invalidOptions.JoinToString(Environment.NewLine)}
-
-Option names must be at least 2 characters long to avoid confusion with short names.
-If you intended to set the short name instead, use the attribute overload that accepts a char.";
-
-            return new TypinException(message.Trim());
-        }
-
         public static TypinException OptionsWithSameName(CommandSchema command, string name, IReadOnlyList<CommandOptionSchema> invalidOptions)
         {
-            var message = $@"
+            string message = $@"
 Command '{command.Type.FullName}' is invalid because it contains {invalidOptions.Count} options with the same name ('{name}'):
 {invalidOptions.JoinToString(Environment.NewLine)}
 
@@ -235,7 +137,7 @@ Names are not case-sensitive.";
 
         public static TypinException OptionsWithSameShortName(CommandSchema command, char shortName, IReadOnlyList<CommandOptionSchema> invalidOptions)
         {
-            var message = $@"
+            string message = $@"
 Command '{command.Type.FullName}' is invalid because it contains {invalidOptions.Count} options with the same short name ('{shortName}'):
 {invalidOptions.JoinToString(Environment.NewLine)}
 
@@ -247,7 +149,7 @@ Short names are case-sensitive (i.e. 'a' and 'A' are different short names).";
 
         public static TypinException OptionsWithSameEnvironmentVariableName(CommandSchema command, string environmentVariableName, IReadOnlyList<CommandOptionSchema> invalidOptions)
         {
-            var message = $@"
+            string message = $@"
 Command '{command.Type.FullName}' is invalid because it contains {invalidOptions.Count} options with the same fallback environment variable name ('{environmentVariableName}'):
 {invalidOptions.JoinToString(Environment.NewLine)}