+++ src/Typin/Typin/Internal/Schemas/CommandParameterSchemaResolver.cs
@@ -2,6 +2,7 @@
 {
     using System.Reflection;
     using Typin.Attributes;
+    using Typin.Internal.Extensions;
     using Typin.Schemas;
 
     /// <summary>
@@ -18,7 +19,7 @@
             if (attribute is null)
                 return null;
 
-            string name = attribute.Name ?? property.Name.ToLowerInvariant();
+            string name = attribute.Name ?? property.Name.ToHyphenCase();
 
             return new CommandParameterSchema(
                 property,